pyproject.toml
[project]
name = "medical-rag-assistant"
version = "0.1.0"
description = ""
authors = [
    {name = "Constanza Florio",email = "florioconstanza@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "langchain (>=1.2.13,<2.0.0)",
    "langchain-groq (>=1.1.2,<2.0.0)",
    "langchain-huggingface (>=1.2.1,<2.0.0)",
    "langchain-chroma (>=1.1.0,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "sentence-transformers (>=5.3.0,<6.0.0)",
    "chromadb (>=1.5.5,<2.0.0)",
    "pymupdf (>=1.27.2.2,<2.0.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "langchain-text-splitters (>=1.1.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false